Move entry api tests to a dedicated package
